Partition the data objects into their own module
Also, remove graph functionality. It was just a bit too unstable
for now. Plan is to do everything MLP first, then move on to CNNs,
and possibly explore graphs later.
