--- crescendo/data.py
@@ -1,515 +1,0 @@
-"""Container for various LightningDataModules."""
-
-from functools import cached_property
-from pathlib import Path
-import pickle
-from tempfile import TemporaryDirectory
-
-from lightning import LightningDataModule
-import numpy as np
-from rich.console import Console
-from sklearn.preprocessing import StandardScaler
-import torch
-from torch.utils.data import DataLoader, TensorDataset
-from torch.utils.data import IterableDataset
-
-from crescendo.utils.datasets import download_california_housing_data
-from crescendo.utils.other_utils import read_json
-
-
-console = Console()
-
-
-def check_batch_size(L_set, batch_size):
-    """There are some bugs (https://stackoverflow.com/a/49035538) that occur
-    when you attempt to use set sizes that are (significantly?) smaller than
-    the requrested batch size. This function checks for that.
-
-    Parameters
-    ----------
-    L_set : int
-        The length of the set provided.
-    batch_size : int
-        The batch size
-    """
-
-    if L_set < batch_size:
-        raise ValueError(
-            "The set size is smaller than the batch size. This is known to "
-            "cause weird problems (https://stackoverflow.com/a/49035538). "
-            f"Please set the batch size to at most {L_set} or get more data."
-        )
-
-
-class XYArrayPropertyMixin:
-    def _apply_ensemble_split(self, data):
-        if self.hparams.ensemble_split_index is None:
-            return data
-        s = f"split-{self.hparams.ensemble_split_index}"
-        split = self.ensemble_splits[s]
-        new_dat = data[split, :]
-        console.log(
-            f"Training data downsampled from {data.shape} -> {new_dat.shape}",
-            style="bold yellow",
-        )
-        return new_dat
-
-    def _apply_feature_selection_logic(self, data):
-        """Runs the feature selection logic on the data based on the provided
-        value of self.hparams.feature_select."""
-
-        if self.hparams.feature_select is None:
-            return data
-
-        # First, split by the comma separation
-        split = self.hparams.feature_select.split(",")
-
-        # For each entry, we split by the second delimiter, a :
-        split = [np.array(xx.split(":")).astype(int).tolist() for xx in split]
-
-        # Edge case
-        if len(split) == 1:
-            return data[:, split[0][0] : split[0][1]]
-
-        # Otherwise we iterate through the splits and concatenate
-        return np.concatenate([data[:, s[0] : s[1]] for s in split], axis=1)
-
-    def _load_data(self, property_name):
-        # Attempt to load from disk
-        fname = f"{property_name}.npy"
-        if self.hparams.data_dir is not None:
-            return np.load(Path(self.hparams.data_dir) / fname)
-
-        # If appropriate file does not exist, attempt to access the internal
-        # value of the object
-        attr_name = f"_{property_name}"
-        dat = getattr(self, attr_name)
-        if dat is not None:
-            return dat
-
-        # Otherwise raise an error
-        raise ValueError(f"_{attr_name} not initialized and dne on disk")
-
-    @cached_property
-    def ensemble_splits(self):
-        path = Path(self.hparams.data_dir) / "splits.json"
-        if not path.exists():
-            return None
-        return read_json(path)
-
-    @cached_property
-    def X_train(self):
-        dat = self._load_data("X_train")
-        dat = self._apply_ensemble_split(dat)
-        return self._apply_feature_selection_logic(dat)
-
-    @cached_property
-    def X_val(self):
-        dat = self._load_data("X_val")
-        return self._apply_feature_selection_logic(dat)
-
-    @cached_property
-    def X_test(self):
-        dat = self._load_data("X_test")
-        return self._apply_feature_selection_logic(dat)
-
-    @cached_property
-    def Y_train(self):
-        dat = self._load_data("Y_train")
-        return self._apply_ensemble_split(dat)
-
-    @cached_property
-    def Y_val(self):
-        return self._load_data("Y_val")
-
-    @cached_property
-    def Y_test(self):
-        return self._load_data("Y_test")
-
-    @cached_property
-    def n_features(self):
-        return self.X_train.shape[1]
-
-    @cached_property
-    def n_targets(self):
-        return self.Y_train.shape[1]
-
-
-class ScaleXMixin:
-    @cached_property
-    def X_train_scaled(self):
-        if self._X_scaler is None:
-            raise ValueError("X-scaler is disabled")
-        return self._X_scaler.transform(self.X_train)
-
-    @cached_property
-    def X_val_scaled(self):
-        if self._X_scaler is None:
-            raise ValueError("X-scaler is disabled")
-        return self._X_scaler.transform(self.X_val)
-
-    @cached_property
-    def X_test_scaled(self):
-        if self._X_scaler is None:
-            raise ValueError("X-scaler is disabled")
-        return self._X_scaler.transform(self.X_test)
-
-    def _setup_X_scaler(self):
-        """
-
-        Parameters
-        ----------
-        stage : Optional[str], optional
-            Description
-        """
-
-        if self.hparams.normalize_inputs:
-            self._X_scaler = StandardScaler()
-            self._X_scaler.fit(self.X_train)
-        else:
-            self._X_scaler = None
-
-
-class DataLoaderMixin:
-    def train_dataloader(self):
-        X = self.X_train.copy()
-        check_batch_size(len(X), self.hparams.dataloader_kwargs["batch_size"])
-        if self._X_scaler is not None:
-            X = self._X_scaler.transform(X)
-        X = torch.tensor(X).float()
-        Y = torch.tensor(self.Y_train.copy()).float()
-        return DataLoader(
-            TensorDataset(X, Y), **self.hparams.dataloader_kwargs
-        )
-
-    def val_dataloader(self):
-        X = self.X_val.copy()
-        check_batch_size(len(X), self.hparams.dataloader_kwargs["batch_size"])
-        if self._X_scaler is not None:
-            X = self._X_scaler.transform(X)
-        X = torch.tensor(X).float()
-        Y = torch.tensor(self.Y_val.copy()).float()
-        return DataLoader(
-            TensorDataset(X, Y), **self.hparams.dataloader_kwargs
-        )
-
-    def test_dataloader(self):
-        X = self.X_test.copy()
-        check_batch_size(len(X), self.hparams.dataloader_kwargs["batch_size"])
-        if self._X_scaler is not None:
-            X = self._X_scaler.transform(X)
-        X = torch.tensor(X).float()
-        Y = torch.tensor(self.Y_test.copy()).float()
-        return DataLoader(
-            TensorDataset(X, Y), **self.hparams.dataloader_kwargs
-        )
-
-
-class ArrayRegressionDataModule(
-    XYArrayPropertyMixin, ScaleXMixin, DataLoaderMixin, LightningDataModule
-):
-    """A standard data module for array data.
-
-    Notes
-    -----
-    ``LightningDataModule`` implements the following methods, which we should
-    keep in the same order:
-
-    - prepare_data: things done on a single compute unit cpu/gpu, such as
-      downloading data, preprocessing, etc.
-    - setup: things done on every ddp process
-    - train_dataloader
-    - val_dataloader
-    - test_dataloader
-    - teardown: called on every process
-
-    See the documentation for further reference
-    https://lightning.ai/docs/pytorch/latest/data/datamodule.html
-
-    Properties
-    ----------
-    data_dir : os.PathLike
-        The location of the data to load. This data is assumed to be kept in
-        files like ``X_train.npy``, ``Y_val.npy``, etc.
-    normalize_inputs : bool
-        If True, will standard-scale the input features to the standard normal
-        distribution using the training set, and execute that transform on the
-        other splits. These objects are stored in ``X_train_scaled``, etc.
-    ensemble_split_index : int, optional
-        If not None, this is an integer referencing the file splits.json in the
-        same directory as the data. This json file contains keys such as
-        "split-0", "split-1", etc., with values corresponding to the training
-        set indexes of the split. You can use
-        ``crescendo.preprocess.array:ensemble_split`` to create this file.
-    feature_select : str, optional
-        If not None, this argument provides some custom functionality to select
-        only a subset of the features provided in the data. For example,
-        ``feature_select="0:200,400:600"`` will select features 0 through 199,
-        inclusive, and 400 through 599, inclusive.
-    dataloader_kwargs : dict, optional
-        A dictionary containing the keyword arguments to pass to all
-        dataloaders.
-    """
-
-    def __init__(
-        self,
-        data_dir,
-        normalize_inputs,
-        ensemble_split_index,
-        feature_select,
-        dataloader_kwargs,
-    ):
-        super().__init__()
-        self.save_hyperparameters(logger=False)
-        self._setup_X_scaler()
-
-
-class CaliforniaHousingDataset(
-    XYArrayPropertyMixin, ScaleXMixin, DataLoaderMixin, LightningDataModule
-):
-    def __init__(
-        self,
-        normalize_inputs,
-        dataloader_kwargs,
-    ):
-        super().__init__()
-        self.hparams.ensemble_split_index = None
-        self.hparams.data_dir = None
-        self.hparams.feature_select = None
-        self.save_hyperparameters(logger=False)
-        with TemporaryDirectory() as t:
-            path = t / Path("california_housing_data")
-            download_california_housing_data(path)
-            self._X_train = np.load(path / "X_train.npy")
-            self._X_val = np.load(path / "X_val.npy")
-            self._Y_train = np.load(path / "Y_train.npy")
-            self._Y_val = np.load(path / "Y_val.npy")
-            self._X_test = None
-            self._Y_test = None
-        self._setup_X_scaler()
-
-
-class GraphInputTargetPredictorMixin:
-    def _apply_ensemble_split(self, data):
-        """The data provided to this ensemble splitting function should be
-        a list of objects, not an array
-
-        Parameters
-        ----------
-        data : list
-
-        Returns
-        -------
-        list
-        """
-
-        if self.hparams.ensemble_split_index is None:
-            return data
-        s = f"split-{self.hparams.ensemble_split_index}"
-        split = self.ensemble_splits[s]
-        new_dat = [data[ii] for ii in split]
-        console.log(
-            f"Training data downsampled from {len(data)} -> {len(new_dat)}",
-            style="bold yellow",
-        )
-        return new_dat
-
-    def _load_data(self, property_name):
-        # Attempt to load from disk
-
-        assert property_name in [
-            "X_train",
-            "X_val",
-            "X_test",
-            "Y_train",
-            "Y_val",
-            "Y_test",
-        ]
-
-        if self.hparams.data_dir is not None:
-            p = Path(self.hparams.data_dir)
-
-            # These are graphs and are stored in pickle format
-            # TODO: must be a better format to store this data in...?
-            if "X" in property_name:
-                return pickle.load(open(p / f"{property_name}.pkl", "rb"))
-
-            # These are vectors in numpy format
-            else:
-                return np.load(p / f"{property_name}.npy")
-
-        # If appropriate file does not exist, attempt to access the internal
-        # value of the object
-        attr_name = f"_{property_name}"
-        dat = getattr(self, attr_name)
-        if dat is not None:
-            return dat
-
-        # Otherwise raise an error
-        raise ValueError(
-            f"_{attr_name} not initialized and does not exist on disk"
-        )
-
-    @cached_property
-    def ensemble_splits(self):
-        path = Path(self.hparams.data_dir) / "splits.json"
-        if not path.exists():
-            return None
-        return read_json(path)
-
-    @cached_property
-    def X_train(self):
-        dat = self._load_data("X_train")
-        return self._apply_ensemble_split(dat)
-
-    @cached_property
-    def X_test(self):
-        return self._load_data("X_test")
-
-    @cached_property
-    def X_val(self):
-        return self._load_data("X_val")
-
-    @cached_property
-    def Y_train(self):
-        dat = self._load_data("Y_train")
-        return self._apply_ensemble_split(dat)
-
-    @cached_property
-    def Y_test(self):
-        return self._load_data("Y_test")
-
-    @cached_property
-    def Y_val(self):
-        return self._load_data("Y_val")
-
-    @cached_property
-    def node_in_feats(self):
-        graph = self.X_train[0]
-        return graph.ndata["features"].shape[1]
-
-    @cached_property
-    def edge_in_feats(self):
-        graph = self.X_train[0]
-        return graph.edata["features"].shape[1]
-
-    @cached_property
-    def n_targets(self):
-        return self.Y_train.shape[1]
-
-
-def collating_function_graph_to_vector(batch):
-    """Collates cd / graph-fixed length vector combination. Recall that
-    in this case, each element of batch is a three vector containing the
-    graph and the target"""
-
-    # Each target is the same length, so we can use standard batching for
-    # it.
-    targets = torch.FloatTensor([xx[1] for xx in batch])
-
-    # However, graphs are not of the same "length" (diagonally on the
-    # adjacency matrix), so we need to be careful. Usually, dgl's batch
-    # method would work just fine here, but for multi-gpu training, we
-    # need to catch some subtleties, since the batch itself is split apart
-    # equally onto many GPU's, but torch doesn't know how to properly split
-    # a batch of graphs. So, we manually partition the graphs here, and
-    # will batch the output of the collating function before training.
-    # This is now just a list of graphs.
-    graphs = [xx[0] for xx in batch]
-
-    return (graphs, targets)
-
-
-class GraphData(IterableDataset):
-    def __init__(self, L):
-        self.L = L
-
-    def __iter__(self):
-        for ii in range(len(self.L)):
-            yield self.L[ii]
-
-
-class GraphToArrayRegressionDataModule(
-    GraphInputTargetPredictorMixin, LightningDataModule
-):
-    """A standard data module for graph-to-array data.
-
-    Notes
-    -----
-    ``LightningDataModule`` implements the following methods, which we should
-    keep in the same order:
-
-    - prepare_data: things done on a single compute unit cpu/gpu, such as
-      downloading data, preprocessing, etc.
-    - setup: things done on every ddp process
-    - train_dataloader
-    - val_dataloader
-    - test_dataloader
-    - teardown: called on every process
-
-    Properties
-    ----------
-    data_dir : os.PathLike
-        The location of the data to load. This data is assumed to be kept in
-        files like ``X_train.npy``, ``Y_val.npy``, etc.
-    normalize_inputs : bool
-        If True, will standard-scale the input features to the standard normal
-        distribution using the training set, and execute that transform on the
-        other splits. These objects are stored in ``X_train_scaled``, etc.
-    ensemble_split_index : int, optional
-        If not None, this is an integer referencing the file splits.json in the
-        same directory as the data. This json file contains keys such as
-        "split-0", "split-1", etc., with values corresponding to the training
-        set indexes of the split. You can use
-        ``crescendo.preprocess.array:ensemble_split`` to create this file.
-    feature_select : str, optional
-        If not None, this argument provides some custom functionality to select
-        only a subset of the features provided in the data. For example,
-        ``feature_select="0:200,400:600"`` will select features 0 through 199,
-        inclusive, and 400 through 599, inclusive.
-    dataloader_kwargs : dict, optional
-        A dictionary containing the keyword arguments to pass to all
-        dataloaders.
-    """
-
-    def __init__(
-        self,
-        data_dir,
-        ensemble_split_index,
-        dataloader_kwargs,
-    ):
-        super().__init__()
-        self.save_hyperparameters(logger=False)
-
-    def train_dataloader(self):
-        X = self.X_train
-        check_batch_size(len(X), self.hparams.dataloader_kwargs["batch_size"])
-        Y = self.Y_train
-        L = [(x, y.tolist()) for x, y in zip(X, Y)]
-        return DataLoader(
-            GraphData(L),
-            collate_fn=collating_function_graph_to_vector,
-            **self.hparams.dataloader_kwargs,
-        )
-
-    def val_dataloader(self):
-        X = self.X_val
-        check_batch_size(len(X), self.hparams.dataloader_kwargs["batch_size"])
-        Y = self.Y_val
-        L = [(x, y.tolist()) for x, y in zip(X, Y)]
-        return DataLoader(
-            GraphData(L),
-            collate_fn=collating_function_graph_to_vector,
-            **self.hparams.dataloader_kwargs,
-        )
-
-    def test_dataloader(self):
-        X = self.X_test
-        check_batch_size(len(X), self.hparams.dataloader_kwargs["batch_size"])
-        Y = self.Y_test
-        L = [(x, y.tolist()) for x, y in zip(X, Y)]
-        return DataLoader(
-            GraphData(L),
-            collate_fn=collating_function_graph_to_vector,
-            **self.hparams.dataloader_kwargs,
-        )
